Add the skeleton for the skimmer.
NanoSkimmer defines the event loop and which producer will be run.
Each producer inherits from the baseproducer as shown by the
TestProducer which is an empty producer and serves as a template.
The executable is defined by the NanoWrapper in bin which calls the
NanoSkimmer class to run the loop.
The produceSkim script is a wrapper for the NanoWrapper and can be used
to submit jobs via HTC.