Introduce the product class for passing previously calculated variables
The Susy1LeptonProduct is a class that can be used to store previously
calculated variables and reuse them in a later producer. This way one
can split the code into smaller chunks and avoid too large producers.
Todo: Maybe use this class directly to store values in the tree instead
of requiring additional memory for each producer specific variable.