Store systematic shifts of JEC and JER in seperate trees
This was a massive change. Systematics which require shift of variables
are now stored in seperate trees so that acceptance of additional jets
or calculation of some high level variables are done with the shifted
variables too.
On the other hand, systematics which only shift normalisation such as
scale factors are only stored in the nominal tree.