Switch to Ultra Legacy Samples
New list of UL samples for 2016 available at data/samples/UltraLegacy.
Keep in mind that the xSection is not final, some samples did not
properly match the sample name to the name listed in xsdb. This need to
be checked.
Technically, everything seems to work just fine but one has to
investigate and see if changes to the code are required.