Redesigning the skimmer,can only process muons/ele
The skimmer is now designed differently.
There is no "NanoSkimmer" class since this will only run on nanoaod.
Therefore the creation of outputfiles and intialisation of all producers
are done in the bin/main.cc.
There is a new datareader class which handles reading from root files by
using TLeafs as this is a very simple approach.
The Susy1LeptonProduct now stores all revelant ouptut information per
event.
It uses std::arrays because this reduces overhead compared to std::vectors.

For now only the Muon and Electron producer work.
TODO:
 - implement scalefactors for muons/electrons
 - Fix the other Producers
 - Finalize the data/config/config.json and define sensible channels
 that are easy to work with and fast to process.